Create primeseries,py
Write a program named primeseries.py, where you will define a function that returns a list of prime
numbers up to a certain max value. A prime number is a number that is divisible by only 1 and itself. In the
program main entry point, use the first argument to the program (sys.argv[1]) as the max value and print
out the prime number list. You can assume that the first argument is always a number greater than 1.
Two test examples:
$ python3 primeseries.py 10
[2, 3, 5, 7]
$ python3 primeseries.py 100
[2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53, 59, 61, 67, 71, 73, 79, 83, 89, 97]