Create primeiter.py
Write a program named primeiter.py, where you will rewrite the 1st question (primeseries) using an
Iterator. Examples of Iterators can be found in the lecture slides. Basically, you will need to implement the
__iter__() and __next__() methods for an Iterator. In the main entry point, use the implemented Iterator in
a for loop to generate the prime numbers up to the first argument to the program. The output should be the
same as the 1st question.