Create primegene.py
Write a program named primegene.py, where you will rewrite the 1st question (primeseries) using a
Generator. Examples of Generators can be found in the lecture slides. Basically, you will need to
implement a function that uses the yield statement. In the main entry point, use the implemented Generator
in a for loop to generate the prime numbers up to the first argument to the program. The output should be
the same as the 1st question.